Fix icon_master.namelist for test case restarts_present
The test case `restarts_present` tests if the restart file is produced icon.
With `lrestart` being true it looks for a restart file on start, so the test
would fail with icon as no restart file is present. Also the restart interval
has to be set to 30s to produce the restart in the outputs.